requirements.txt
Flask==2.3.3
pandas==2.1.1
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
openpyxl==3.1.2
xlrd==2.0.2
Werkzeug==2.3.7